backend/backup_20260108/requirements.txt
# GeminiCLI Restaurant Pipeline Dependencies
requests
python-dotenv
google-api-python-client
openai
youtube-transcript-api
google-generativeai
